feat(modal): implement modal context and hooks for managing shown modals
