+++ src/components/modal/main.tsx
@@ -1,42 +1,18 @@
 import { Global } from "@emotion/react";
 import { IconX } from "@tabler/icons-react";
 import { AnimatePresence, motion, useReducedMotion } from "framer-motion";
-import { useEffect, useId, useMemo, useState } from "react";
+import { ReactNode, useEffect, useId, useMemo, useState } from "react";
 import { createPortal } from "react-dom";
 import { scrollbarsInvisible } from "../styles.ts";
 import { AnimatedTooltip } from "../../modules/button";
-import { create } from "zustand";
-import { produce } from "immer";
 import { ZIndexProvider } from "../../modules/zindex/zindex";
 import IconLoader from "../../modules/loader";
-
-interface ShownModalStore {
-  shownModals: string[];
-  nextId: number;
-
-  add(id: string): void;
-  remove(id: string): void;
-  getNewId(): number;
-}
-
-const useShownModals = create<ShownModalStore>()((set, get) => ({
-  shownModals: [],
-  nextId: 0,
-
-  add: id => set(prev => produce(prev, prev => {
-    prev.shownModals.push(id)
-  })),
-  remove: id => set(prev => produce(prev, prev => {
-    prev.shownModals = prev.shownModals.filter(x => x != id)
-  })),
-  getNewId: () => (set(prev => produce(prev, prev => {
-    prev.nextId++
-  })), get().nextId - 1),
-}));
+import { ModalBlurContext, useShownModals } from "./context";
 
 export default function Modal({
   title,
   children,
+  containerRef,
   shown,
   onClose,
   rootId = "root",
@@ -47,9 +23,11 @@ export default function Modal({
   withUi = true,
   withCloseButton = true,
   className,
+  position = "center",
 }: {
   title?: string;
-  children?: React.ReactNode;
+  children?: ReactNode;
+  containerRef?: React.Ref<HTMLDivElement>;
   shown: boolean;
   onClose?: () => void;
   rootId?: string;
@@ -60,6 +38,7 @@ export default function Modal({
   withUi?: boolean;
   withCloseButton?: boolean;
   className?: string;
+  position?: "center" | "bottom";
 }) {
   const id = useId();
   const modals = useShownModals(v => v.shownModals);
@@ -159,7 +138,7 @@ export default function Modal({
           inset: 0,
           overflow: "hidden",
           display: "grid",
-          placeItems: "center",
+          placeItems: position == "center" ? "center" : "end center",
         }}
         id={`modal-content-${id}`}
         aria-hidden={!shown}
@@ -167,6 +146,7 @@ export default function Modal({
       >
         <div
           className={className}
+          ref={containerRef}
           css={[{
             display: "flex",
             flexDirection: "column",
@@ -202,7 +182,9 @@ export default function Modal({
           </div>}
 
           <ZIndexProvider min={zIndex + 1} max={zIndex + 300}>
-            {children}
+            <ModalBlurContext.Provider value={lastShown != id}>
+              {children}
+            </ModalBlurContext.Provider>
           </ZIndexProvider>
         </div>
       </motion.div>}
